Split utils key and JSON helpers
